Check for newlines in the CSV entries
Newlines inside an entry from a CSV row will not prevent this script
from generating the type_nodes setups, but will hard crash Flame if you
to try to load the resulting setups.

Due to newlines being the delimiters in a CSV file, any newlines inside
the row entries will be treated as invalid CSV data and an error message
will be printed.