Change FlameLineEditFileBrowse default path
Have FlameLineEditFileBrowse open at Flame setups directory instead of
user folder if line edit path is not valid.  This change is because it
is intended that the user is generating slate setups inside the setups
directory, not their user folder.  Make starting path optional
since it will not be set in the View, but later in the Controller.